lambda_function/requirements.txt
langchain>=0.2.0
langchain-openai>=0.0.3
pyyaml>=6.0.1
python-dotenv>=1.0.1
boto3>=1.34.0 
